Revert "지도 컴포넌트 분리"
This reverts commit 26f1b7272b3b9ba0bb18a82eac8930ce4ff7e8f4.
