+++ src/app/map/MapApp.tsx
@@ -15,11 +15,8 @@ import useGetLocations from "./hooks/useGetLocations"
 
 import useKakaoApi from "./../../components/api/useKakaoApi"
 
-import { useStationSetting } from "./hooks/useStationSetting"
-// import MapMarkerOverlay from "./MapMarkerOverlay"
+import MapMarkerOverlay from "./MapMarkerOverlay"
 import MapClickHandler from "./MapClickHandler"
-import MapControls from "./MapControls"
-import MapBoundaryInfo from "./MapBoundaryInfo"
 
 interface MapAppProps {
   setSelectedStation: React.Dispatch<React.SetStateAction<string | null>>
@@ -42,21 +39,14 @@ const MapApp: React.FC<MapAppProps> = ({ setSelectedStation }) => {
 
   const { loading: apiLoading, error: apiError } = useKakaoApi()
 
+  const currentNearestStation = useNearStation(position, locations)
   const { map, setMap, containerRef } = useMapResize()
-  // const currentNearestStation = useNearStation(position, locations)
 
-  // const [initNearestStation, setInitNearestStation] = useState<
-  //   typeof currentNearestStation | null
-  // >(null)
+  const [initNearestStation, setInitNearestStation] = useState<
+    typeof currentNearestStation | null
+  >(null)
 
-  // const displayStation = initNearestStation || currentNearestStation
-
-  const { displayStation, initNearestStation, currentNearestStation } =
-    useStationSetting({
-      position,
-      locations,
-      setSelectedStation,
-    })
+  const displayStation = initNearestStation || currentNearestStation
 
   // const nearestStation = useNearStation(currentNearestStation, locations)
 
@@ -70,64 +60,41 @@ const MapApp: React.FC<MapAppProps> = ({ setSelectedStation }) => {
   //   },
   // ]
 
-  // useEffect(() => {
-  //   if (
-  //     !initNearestStation &&
-  //     currentNearestStation &&
-  //     currentNearestStation.title
-  //   ) {
-  //     setInitNearestStation(currentNearestStation)
-  //     setSelectedStation(currentNearestStation.title)
-  //     console.log("초기 가장 가까운 측정소:", currentNearestStation.title)
-  //   }
-  // }, [initNearestStation, currentNearestStation, setSelectedStation])
+  useEffect(() => {
+    if (
+      !initNearestStation &&
+      currentNearestStation &&
+      currentNearestStation.title
+    ) {
+      setInitNearestStation(currentNearestStation)
+      setSelectedStation(currentNearestStation.title)
+      console.log("초기 가장 가까운 측정소:", currentNearestStation.title)
+    }
+  }, [initNearestStation, currentNearestStation, setSelectedStation])
 
-  // useEffect(() => {
-  //   // 사용자가 지도를 클릭해서 선택한 측정소가 있으면 변경하지 않습니다.
-  //   if (!displayStation?.title) return
-  //   setSelectedStation(displayStation.title)
-  //   console.log("가장 가까운 측정소:", displayStation.title)
-  // }, [displayStation, setSelectedStation])
+  useEffect(() => {
+    // 사용자가 지도를 클릭해서 선택한 측정소가 있으면 변경하지 않습니다.
+    if (!displayStation?.title) return
+    setSelectedStation(displayStation.title)
+    console.log("가장 가까운 측정소:", displayStation.title)
+  }, [displayStation, setSelectedStation])
 
   // const visibleMarkers = useVisibleMarkers(locations, bounds)
 
   if (apiLoading || locationLoading) return <div>로딩중...</div>
   if (apiError) return <div>카카오맵 API 로딩 실패: {apiError.message}</div>
 
-  const handleBackToCurrentLocation = () => {
-    console.log("현재 위치로 이동")
-    refetch()
-
-    if (map && position) {
-      map.panTo(
-        new (window as any).kakao.maps.LatLng(position.lat, position.lng)
-      )
-    }
-  }
-
-  const targetStation =
-    initNearestStation?.title || currentNearestStation?.title
-  if (targetStation) {
-    setSelectedStation(targetStation)
-    console.log("가장 가까운 측정소:", targetStation)
-  }
-
   return (
     // Map 내부에서 loading 상태를 관찰하고 있기 때문에 conditional rendering를 하지 않아도 됩니다.
     <>
-      {/* <div style={{ margin: "0.75em 0" }}>
+      <div style={{ margin: "0.75em 0" }}>
         <strong>현재 위치: </strong> {address}
         {displayStation && (
           <p>
             <strong>가장 가까운 측정소: </strong> {displayStation.title}
           </p>
         )}
-      </div> */}
-      <MapControls
-        address={address}
-        displayStation={displayStation}
-        onBackToCurrentLocation={handleBackToCurrentLocation}
-      />
+      </div>
       <div ref={containerRef}>
         <Map
           center={position}
@@ -169,7 +136,7 @@ const MapApp: React.FC<MapAppProps> = ({ setSelectedStation }) => {
           />
         </Map>
       </div>
-      {/* {bounds && (
+      {bounds && (
         <div style={{ marginTop: "1em" }}>
           <strong>지도 경계:</strong>
           <br />
@@ -216,8 +183,7 @@ const MapApp: React.FC<MapAppProps> = ({ setSelectedStation }) => {
         >
           현재 위치로 돌아가기
         </button>
-      </div> */}
-      <MapBoundaryInfo bounds={bounds} />
+      </div>
     </>
   )
 }
